Fixed some warnings during compilation

+++ stm32/lib/storage/include/fifo.h
@@ -91,7 +91,7 @@ FramStatus FramPut(const uint8_t *data, uint16_t num_bytes);
  * @param    len Length of data
  * @return   See FramStatus
  */
-FramStatus FramGet(uint8_t *data, uint8_t *len);
+FramStatus FramGet(uint8_t *data, size_t *len);
 
 /**
  * @brief Get the current number of measurements stored in the buffer
